% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connections.R
\name{make_connections}
\alias{make_connections}
\title{Create a dataframe listing nodes and details of how they connect to other nodes}
\usage{
make_connections(data, node_col, connect_col)
}
\arguments{
\item{data}{Dataframe holding the information you'd like to reformat}

\item{node_col}{The \code{data} column containing identifiers for nodes}

\item{connect_col}{The \code{data} column containing the ids of objects connecting nodes, for example, project IDs, locations, mentees}
}
\value{
A dataframe containign details of connections between nodes via 'connecting objects'.
}
\description{
Creates a df of unique nodes and vertices combinations. For example, it might list all team members/nodes working on projects, connected via a project ID.
}
