% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{make_edges}
\alias{make_edges}
\title{Calculate vertices for network visualisations.}
\usage{
make_edges(
  connections_data,
  node_id,
  node_number,
  connection_id,
  join_field = deparse(substitute(connection_id))
)
}
\arguments{
\item{connections_data}{data detailing connections between nodes}

\item{node_id}{field which identifies nodes}

\item{node_number}{field identifying number for each node}

\item{connection_id}{field identifying connecting objects}

\item{join_field}{field to join connections data back onto itself}
}
\value{
A dataframe containing details of vertices between nodes, including weights.
}
\description{
Calculates weightings between nodes for a network diagram.
}
