% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nodes.R
\name{make_nodes}
\alias{make_nodes}
\title{Reformat dataframes in an appropriate format for network analysis.}
\usage{
make_nodes(
  connection_data,
  connection_node_col,
  attribute_data,
  attribute_node_col,
  ... = NULL
)
}
\arguments{
\item{connection_data}{Dataframe holding the information you'd like to reformat}

\item{connection_node_col}{The field containing the node ids/values in \code{connection_data}}

\item{attribute_data}{Dataframe containing node attribute data}

\item{attribute_node_col}{Column name in \code{attribute_data} that contains the unique common node identifier}

\item{...}{Attribute variables in \code{attribute_data}}
}
\value{
A df of nodes as per \code{node_col}
}
\description{
Take a data frame and create a df of unique nodes containing their df unique ids, and a 'node id' as well.
}
