% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nodes.R
\name{make_nodes}
\alias{make_nodes}
\title{Reformat dataframes in an appropriate format for network analysis.}
\usage{
make_nodes(attribute_data, node_col, ...)
}
\arguments{
\item{attribute_data}{Dataframe containing node attribute data}

\item{node_col}{Column name in \code{attribute_data} that contains the unique common node identifier}

\item{...}{Attribute variables in \code{attribute_data}. Note the function assumes there is only one value for each attribute per node.}
}
\value{
A dataframe of nodes as per \code{node_col}
}
\description{
Take a data frame and create a dataframe of unique nodes containing their dataframe unique ids, and a 'node id' as well.
}
